@@ -1,12 +1,13 @@
 package com.faob.motionlayoutcarousel
 
-import android.support.constraint.ConstraintLayout
-import android.support.v7.widget.RecyclerView
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
 
-class DemoAdapter(private val demoList: ArrayList<DemoAdapter.Demo>) : RecyclerView.Adapter<DemoAdapter.ViewHolder>(){
+class DemoAdapter(private val demoList: ArrayList<Demo>) :
+    RecyclerView.Adapter<DemoAdapter.ViewHolder>() {
 
     data class Demo(val item: String, val layoutId: Int)
 
@@ -24,7 +25,8 @@ class DemoAdapter(private val demoList: ArrayList<DemoAdapter.Demo>) : RecyclerV
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val row: ConstraintLayout = LayoutInflater.from(parent.context).inflate(R.layout.row, parent, false) as ConstraintLayout
+        val row: ConstraintLayout = LayoutInflater.from(parent.context)
+            .inflate(R.layout.row, parent, false) as ConstraintLayout
         return ViewHolder(row);
     }
 
