@@ -2,10 +2,10 @@ package com.faob.motionlayoutcarousel
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 
 class MainActivity : AppCompatActivity() {
 
@@ -15,14 +15,20 @@ class MainActivity : AppCompatActivity() {
 
         val layoutList = arrayListOf(R.layout.demo1, R.layout.demo2)
         val demoList = arrayListOf<DemoAdapter.Demo>()
-        for (i in 0 until layoutList.size){
-            demoList.add(DemoAdapter.Demo("DEMO ${i+1}", layoutList[i]))
+        for (i in 0 until layoutList.size) {
+            demoList.add(DemoAdapter.Demo("DEMO ${i + 1}", layoutList[i]))
         }
 
         val recyclerView = findViewById<RecyclerView>(R.id.rv)
         recyclerView.setHasFixedSize(true)
-        recyclerView.addItemDecoration(DividerItemDecoration(this, LinearLayoutManager.VERTICAL))
-        recyclerView.layoutManager = LinearLayoutManager(this)
+        recyclerView.addItemDecoration(
+            DividerItemDecoration(
+                this,
+                LinearLayoutManager.VERTICAL
+            )
+        )
+        recyclerView.layoutManager =
+            LinearLayoutManager(this)
         recyclerView.adapter = DemoAdapter(demoList)
     }
 
